challenge_level1/challenge2_loop: add for loop
In this test we check three times values loaded from memory.
We need to check the number of values we have checked, written in t5. If
we don't do that we would start reading outside the test data block.

For that we error out if the sum of the two values don't correspond to
the control value in memory.
If the sum is correct, we decrease the t5 and check if we reached the
end of the test wirtten in the test data section. If so we can directly
end the test.