challenge_level2/challenge1: Only execute supported instructions
When trying to compile the assembly code generated by aapg, we get an
error about unrecognized opcode, for example:
divuw, remuw, mulw, etc

This instructions are part of the RV64 M Extension (Integer
Multiplication and Division). Our compiler only supports RV32 and the
hackathon concentrates only on RV32I specifically. To fix the test, we
disable rel_rv64m by setting it to 0. With that we only generate
supported assembly code. Others like float-rounding option do not need
adjustment as the F Extension is not enabled in the
isa-instruction-distribution.